% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge-utils.R
\name{badge2text}
\alias{badge2text}
\title{Convert a badge SVG to text}
\usage{
badge2text(svg_badge_url, full_colour = TRUE)
}
\arguments{
\item{svg_badge_url}{URL of a SVG badge.}

\item{full_colour}{whether to process the badge colours.}
}
\value{
a badge in text form, optionally coloured.
}
\description{
Convert a badge SVG to text
}
