% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_animations.R
\name{create_progressbar}
\alias{create_progressbar}
\alias{update.progressbar}
\alias{render.progressbar}
\title{Creates an animated progress bar}
\usage{
create_progressbar(format = "[[|][|/-\\\\][ ]]", width = 25,
  refresh = 200)

\method{update}{progressbar}(object, progress, ...)

\method{render}{progressbar}(x, ...)
}
\arguments{
\item{format}{Character vector containing format of the resulting progress bar. See 'details' for more information.}

\item{width}{The width in characters of the progress bar.}

\item{refresh}{The refresh rate in milliseconds of the animation.}

\item{object}{A \code{progressbar}.}

\item{progress}{Fraction indicating the progression of the progress bar. Must lie in \code{[0,1]}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{A \code{progressbar}.}
}
\description{
Creates an animated progress bar
}
\examples{
\dontrun{
bar = create_progressbar(format = "[[|][|/-\\\\][ ]]")
n_operations = 1000

for(i in 1:n_operations) {
  bar = update(bar, i/n_operations)
  cat("\\r", render(bar))
  Sys.sleep(0.01)
}}
}
