% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{create_text_table}
\alias{create_text_table}
\title{Creates a text table}
\usage{
create_text_table(string, n_cols = get_square_grid(length(string))$columns)
}
\arguments{
\item{string}{character vector of strings to reformat.}

\item{n_cols}{number of resulting columns.}
}
\value{
A vector of strings per row, forming together a table.
}
\description{
Creates a text table
}
\examples{
cat(create_text_table(LETTERS),sep = "\\n")
}
