% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{createDescImports}
\alias{createDescImports}
\title{Creates a list of necessary imports for the DESCRIPTION file.s}
\usage{
createDescImports()
}
\description{
Creates a list of necessary imports for the DESCRIPTION file.s
}
\examples{
createDescImports()
}
