% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-utils.R
\name{.extract_packages}
\alias{.extract_packages}
\title{Get the package names from file content}
\usage{
.extract_packages(content)
}
\arguments{
\item{content}{The content of an .R file}
}
\value{
a list of package names which were loaded in the content
}
\description{
Get the package names from file content
}
