% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{.get_examples}
\alias{.get_examples}
\title{Displays sample functions for various genres of functions.}
\usage{
.get_examples(collection)
}
\arguments{
\item{collection}{The collection name. See \code{\link{load_package_collection}}.}
}
\description{
Displays sample functions for various genres of functions.
}
