% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{.impls}
\alias{.impls}
\title{Retrieves generic function implementation}
\usage{
.impls(generic)
}
\arguments{
\item{generic}{A string with the name of the generic function.}
}
\value{
A vector of class names for which argument 'generic' provides an implementation.
}
\description{
Retrieves generic function implementation
}
\examples{
.impls('print')
}
