% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{.is_registered}
\alias{.is_registered}
\title{Internal function to check whether the current user has the working directory registered.}
\usage{
.is_registered(verbose = FALSE, return_data = TRUE)
}
\arguments{
\item{verbose}{Whether to print the user name and registered working directory.}

\item{return_data}{Whether to return a boolean or a tibble containing the found user.}
}
\value{
Either a boolean (if return_data is FALSE) or a tibble containing fields 'desc' (computer description),
'usr' (computer user name) and 'location' (working directory base).
}
\description{
Internal function to check whether the current user has the working directory registered.
}
