% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{.value_str}
\alias{.value_str}
\title{Creates a nice string representation of a variable.}
\usage{
.value_str(x, show_class = FALSE)
}
\arguments{
\item{x}{The variable for which a string representation is created.}

\item{show_class}{Whether to show the class of 'x'. Defaults to FALSE.}
}
\value{
A character vector with the string representation of 'x'.
}
\description{
Creates a nice string representation of a variable.
}
\note{
Example: \code{.value_str(c(1,2,3))}
}
