% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{format_duration}
\alias{format_duration}
\title{Format time duration}
\usage{
format_duration(start, end)
}
\arguments{
\item{start, end}{datetime objects as obtained via \code{\link[base]{Sys.time}}}
}
\value{
A string representation of the duration.
}
\description{
Format time duration
}
