% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{generic_implementations}
\alias{generic_implementations}
\title{Retrieves generic function implementation}
\usage{
generic_implementations(generic)
}
\arguments{
\item{generic}{A string with the name of the generic function.}
}
\value{
A vector of class names for which argument 'generic' provides an implementation.
}
\description{
Retrieves generic function implementation
}
\examples{
impls = generic_implementations('plot')
}
