% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-utils.R
\name{get_packages}
\alias{get_packages}
\title{Extract package names from R files in folder}
\usage{
get_packages(folder, recursive = TRUE)
}
\arguments{
\item{folder}{The base folder name, containing the .R files}

\item{recursive}{Should the listing recurse into directories?}
}
\value{
A named list. For each .R file containing package loading, a list of package names
}
\description{
Extract package names from R files in folder
}
