% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{get_survival_estimate}
\alias{get_survival_estimate}
\title{Get estimate of timepoints for a given survival probability}
\usage{
get_survival_estimate(sfit, survival = 0.5)
}
\arguments{
\item{sfit}{A survfit object}

\item{survival}{The survival probability for which a timepoint estimate is needed. Default is 0.5 (median survival)}
}
\value{
A named list or matrix with elements surv (estimate), lower and upper (confidence interval). The attribute "survival" is set
to the argument survival
}
\description{
Get estimate of timepoints for a given survival probability
}
\examples{
fit = cph(Surv(time=time, event = status==2) ~ age + sex, data=lung, x = T, y=T, surv=T)
sfit = survfit(fit)
sfit2 = survfit(fit, newdata=lung[1:20,])

get_survival_estimate(sfit) #get median survival for all data point
get_survival_estimate(sfit2) #get median survival for patients 1-20 seperately
}
