% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{ggplot_breaks}
\alias{ggplot_breaks}
\title{Nice plotting axis breaks}
\usage{
ggplot_breaks(...)
}
\arguments{
\item{...}{Arguments passed on to \code{get_breaks}
\describe{
  \item{limits}{The limits of the axis. May be a vector of 2 elements with lower and upper bounds, or a
single digit (which is the upper bound, the lower bound is then assumed to be 0).}
  \item{N}{The step size. The eventual intervals will be multiples of the divisors of \code{N} or
multiples of \code{N} when \code{multiples_only} is \code{TRUE}. Defaults to 10.}
  \item{max_breaks}{Maximum amount of breaks, defaults to 10.}
  \item{int_only}{Whether only integer divisors of \code{N} may be used for interval sizes, defaults to \code{TRUE}.}
  \item{multiples_only}{Whether only multiples of \code{N} can be used, defaults to \code{FALSE}.}
  \item{include_bounds}{Whether the resulting bounds should include \code{min} and \code{max}. Defaults to \code{TRUE}.}
}}
}
\value{
A \code{\link{get_breaks}} function with filled-in parameters which expects limits.
}
\description{
This makes usage easier in \pkg{ggplot2} as the limits may not be always be known before plotting.
}
\examples{
\dontrun{ggplot() + scale_x_continuous(breaks = ggplot_breaks(N=12, max_breaks=15))}
}
\concept{axis break functions}
