% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils-package.r
\docType{package}
\name{HGUtils-package}
\alias{HGUtils}
\alias{HGUtils-package}
\title{Collection of utility functions}
\description{
A collection of some useful utility functions for plotting, project management and survival.
}
\author{
\strong{Maintainer}: Héctor van den Boorn \email{h.g.vandenboorn@amc.uva.nl}

}
