% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Installs and loads specified packages}
\usage{
install_load_packages(..., install = TRUE)
}
\arguments{
\item{...}{A list of package names}

\item{install}{Whether to install packages when missing, defaults to TRUE}
}
\description{
Installs and loads specified packages
}
\examples{
\dontrun{
install_load_packages(c('ggplot2','dplyr'))
install_load_packages('ggplot2','dplyr',install=FALSE)}
}
\seealso{
Other initialization functions: \code{\link{is_registered}},
  \code{\link{load_common_packages}}, \code{\link{startup}}
}
\concept{initialization functions}
