% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{is_registered}
\alias{is_registered}
\title{Checks whether the current user has the working directory registered.}
\usage{
is_registered(verbose = FALSE)
}
\arguments{
\item{verbose}{Whether to print the user name and registered working directory.}
}
\value{
Boolean indicating whether the user is registered or not.
}
\description{
Checks whether the current user has the working directory registered.
}
\examples{
is_registered()
}
\seealso{
Other initialization functions: \code{\link{install_load_packages}},
  \code{\link{load_common_packages}}, \code{\link{startup}}
}
