% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{is_valid_pkgname}
\alias{is_valid_pkgname}
\title{Validate a package name}
\usage{
is_valid_pkgname(pkg)
}
\arguments{
\item{pkg}{The name of a package. May be a vector or a single string.}
}
\value{
A boolean indicating whether the package name is valid.
}
\description{
Naming rule obtained from \emph{'Writing R Extensions'} manual.
}
\examples{
is_valid_pkgname("hgutils") #valid
is_valid_pkgname("ggplot2") #valid
is_valid_pkgname("pkg2.-1") #invalid
}
\references{
\href{https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#The-DESCRIPTION-file}{'Writing R Extensions'} manual.
}
\seealso{
Other initialization functions: \code{\link{is_registered}},
  \code{\link{startup}}, \code{\link{use_common_packages}},
  \code{\link{use_packages}}
}
\concept{initialization functions}
