% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{load_common_packages}
\alias{load_common_packages}
\title{Loads (and optionally install) commonly used packages}
\usage{
load_common_packages(install = T)
}
\arguments{
\item{install}{Whether to install a package if missing}
}
\description{
Loads (and optionally install) commonly used packages
}
