% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-utils.R
\name{message_colors}
\alias{message_colors}
\alias{.cmess}
\alias{.cmess_bg}
\alias{.cwarn}
\alias{.cwarn_bg}
\alias{.chint}
\alias{.chint_bg}
\alias{.csucc}
\alias{.csucc_bg}
\alias{.cfail}
\alias{.cfail_bg}
\alias{.cnumb}
\alias{.cnumb_bg}
\title{Standardized message colors}
\arguments{
\item{...}{the text to be colored.}
}
\value{
A function to color text
}
\description{
Standardized message colors
}
\section{Functions}{
\itemize{
\item \code{.cmess}: For warning messages.

\item \code{.cmess_bg}: For warning messages.

\item \code{.cwarn}: For warning messages.

\item \code{.cwarn_bg}: For warning messages.

\item \code{.chint}: For warning messages.

\item \code{.chint_bg}: For warning messages.

\item \code{.csucc}: For success messages.

\item \code{.csucc_bg}: For success messages.

\item \code{.cfail}: For fail messages.

\item \code{.cfail_bg}: For fail messages.

\item \code{.cnumb}: For fail messages.

\item \code{.cnumb_bg}: For fail messages.
}}

