% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_genre.R
\docType{data}
\name{pkg_genre}
\alias{pkg_genre}
\title{List of categorized packages}
\format{A data frame with 8 rows and 2 variables:
\describe{
  \item{name}{The name}
  \item{packages}{list of packages}
}}
\usage{
pkg_genre
}
\description{
List of categorized packages
}
\keyword{datasets}
