% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge-utils.R
\name{print.badge}
\alias{print.badge}
\title{Prints coloured badges}
\usage{
\method{print}{badge}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints coloured badges
}
