% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{print.percentage_table}
\alias{print.percentage_table}
\title{Print a formatted percentage table}
\usage{
\method{print}{percentage_table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{percentage_table}}

\item{...}{unused}
}
\description{
Print a formatted percentage table
}
\examples{
print(percentage_table(iris$Species))
}
