% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_animations.R
\name{progressbar}
\alias{progressbar}
\title{Creates an animated progress bar}
\usage{
progressbar(format = "[[|][|/-\\\\][ ]]", width = 25, refresh = 200,
  n_iterations = NULL)
}
\arguments{
\item{format}{Character vector containing format of the resulting progress bar. See 'details' for more information.}

\item{width}{The width in characters of the progress bar.}

\item{refresh}{The refresh rate in milliseconds of the animation.}

\item{n_iterations}{Optional parameter, specifies the number of total iterations. When updating the progress bar it
is then sufficient to specify the current iteration number.}
}
\description{
Creates an animated progress bar
}
\examples{
\dontrun{
#simple progressbar
bar = progressbar(format = "[[|][|/-\\\\][ ]]")
#fancy progressbar using UTF-8 codes
bar2 = progressbar(format="\\u25ba[\\u2589][\\u2580\\u2584][\\u3000]\\u25c4")
n_operations = 1000

for(i in 1:n_operations) {
  bar = update(bar, i/n_operations)
  cat("\\r", render(bar))
  Sys.sleep(0.01)
}}
}
