% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-utils.R
\name{read.description}
\alias{read.description}
\title{Reads the DESCRIPTION file}
\usage{
read.description()
}
\value{
The DESCRIPTION file as a named list.
}
\description{
Reads the DESCRIPTION file
}
\examples{
\dontrun{description = read.description()}
}
