% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{retrieve_ellipsis}
\alias{retrieve_ellipsis}
\title{Update default parameters with ellipsis}
\usage{
retrieve_ellipsis(default, ...)
}
\arguments{
\item{default}{A named list of default values for parameters}

\item{...}{Optional parameters to override the default parameters. A warning is shown if the names do not match.}
}
\value{
The updated list of parameters with possible new values.
}
\description{
Update default parameters with ellipsis
}
\examples{
foo = function(...) {
default = list(a=1)
updated_args = retrieve_ellipsis(default, ...)
}
}
