% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{rmNA}
\alias{rmNA}
\title{Remove \code{NA}}
\usage{
rmNA(vec)
}
\arguments{
\item{vec}{A vector which may contain \code{NA} elements.}
}
\value{
A vector without \code{NA} elements.
}
\description{
Removes any \code{NA} element from a vector.
}
\examples{
rmNA(c(1,NA,5,6))
}
