% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{seperate_values}
\alias{seperate_values}
\title{Seperate values
Seperates real numbers from one another with a minimum distance, bounded by lower and upper values and constraint to be as
close as possible to their original values.}
\usage{
seperate_values(x, distance = 0.05, min = 0, max = 1)
}
\arguments{
\item{x}{A sorted numerical vector of real numbers.}

\item{distance}{The minimum distance between subsequent numbers}

\item{min}{The minimum value of a number}

\item{max}{The maximum value of a number}
}
\value{
A numerical vector with the same length as x, with numbers bounded by min and max, close to their original values and
        with the minimum allowed distance between subsequent values.
}
\description{
Seperate values
Seperates real numbers from one another with a minimum distance, bounded by lower and upper values and constraint to be as
close as possible to their original values.
}
\examples{
seperates_values(c(0.3,0.4,0.41), distance = 0.05, min = 0, max = 1)
}
