% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{set_DESCRIPTION_imports}
\alias{set_DESCRIPTION_imports}
\title{Creates a list of necessary imports for the DESCRIPTION file.s}
\usage{
set_DESCRIPTION_imports()
}
\description{
Creates a list of necessary imports for the DESCRIPTION file.s
}
\examples{
set_DESCRIPTION_imports()
}
