% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{set_package_imports}
\alias{set_package_imports}
\title{Creates a list of necessary imports for the DESCRIPTION file.}
\usage{
set_package_imports(skip_prompt = FALSE)
}
\arguments{
\item{skip_prompt}{Whether to skip the prompt or to edit the DESCRIPTION file directly.}
}
\description{
Creates a list of necessary imports for the DESCRIPTION file.
}
\examples{
\dontrun{set_package_imports(skip_prompt=TRUE)}
}
