% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{sort_version}
\alias{sort_version}
\title{Sorts version numbers}
\usage{
sort_version(X)
}
\arguments{
\item{X}{A vector of version numbers.}
}
\value{
A sorted list of version numbers
}
\description{
Sort version numbers ascending according to \code{\link[utils]{compareVersion}}.
}
\examples{
sort_version(c("3.1.2","3.1","3.1.0"))
}
\seealso{
\code{\link[utils]{compareVersion}}
}
