% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{startup}
\alias{startup}
\title{Sets the working directory automatically to Dropbox folder}
\usage{
startup(folder = "source_webinterface/")
}
\arguments{
\item{folder}{String to specify subfolder name}
}
\description{
Sets the working directory automatically to Dropbox folder
}
\examples{
startup("source_webinterface/")
}
