% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{stfu}
\alias{stfu}
\title{STFU}
\usage{
stfu(expr)
}
\arguments{
\item{expr}{An expression to evaluate in silence.}
}
\value{
Returns (invisibly) the result of \code{expr}.
}
\description{
STFU: \strong{S}top \strong{T}umult \strong{F}rom o\strong{U}tput
}
\section{Warning}{

Make sure to call this function \strong{always} directly on the expression and never indirectly such as via pipes.
Example: \code{stfu(expr)} is correct, but \code{expr \%>\% stfu} will hide the output. However, the \code{expr} argument itself may contain pipes.
}

\examples{
stfu(print("hi"))
}
