% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survEstimateTime.R
\name{survEstimateTime}
\alias{survEstimateTime}
\alias{survEstimateTime.survfit}
\alias{survEstimateTime.list}
\alias{survEstimateTime.coxph}
\title{Survival time estimation}
\usage{
survEstimateTime(fit, survival, ...)

\method{survEstimateTime}{survfit}(sFit, survival = 0.5)

\method{survEstimateTime}{list}(sEst, survival = 0.5)

\method{survEstimateTime}{coxph}(fit, survival = 0.5, newdata = NULL)
}
\arguments{
\item{fit}{A \code{\link[survival]{coxph}} object}

\item{survival}{the survival probability for which the timepoint is estimated.}

\item{...}{unused}

\item{sFit}{A \code{\link[survival]{survfit} object}}

\item{sEst}{A \code{\link[rms]{survest.cph} object}}

\item{newdata}{A dataframe containing predictors for which predictors are desired. See \code{\link[survival]{survfit.coxph}}.}
}
\value{
A named list or matrix with elements surv (estimate), lower and upper (confidence interval). The attribute 'survival' is added to the
result and set to the argument survival probability
}
\description{
Get estimate of timepoints for a given survival probability.
}
\section{Methods (by class)}{
\itemize{
\item \code{survfit}: for survfit objects.

\item \code{list}: for survest objects.

\item \code{coxph}: A \code{\link[survival]{coxph}} object.
}}

\examples{
library(rms)
fit = cph(Surv(time=time, event=status) ~ age + sex, data=lung, x=TRUE, y=TRUE, surv=TRUE)
sFit = survfit(fit)
sFit2 = survfit(fit, newdata=lung[1:20,])

survEstimateTime(sFit) #get median survival for all data points in the dataset.
survEstimateTime(sFit2) #get median survival for patients 1-20 seperately
}
