% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survEstimateTime.R
\name{survEstimateTime.survRes}
\alias{survEstimateTime.survRes}
\title{Survival time estimation for survRes objects}
\usage{
\method{survEstimateTime}{survRes}(sRes, survival = 0.5)
}
\arguments{
\item{sRes}{A \code{sRes} object}

\item{survival}{the survival probability for which the timepoint is estimated.}
}
\description{
Survival time estimation for survRes objects
}
