% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survEstimateTime.R
\name{survTimeEstimate}
\alias{survTimeEstimate}
\alias{survTimeEstimate.survRes}
\alias{survTimeEstimate.survfit}
\alias{survTimeEstimate.list}
\alias{survTimeEstimate.coxph}
\title{Survival time estimation}
\usage{
survTimeEstimate(fit, survival, ...)

\method{survTimeEstimate}{survRes}(fit, survival = 0.5, ...)

\method{survTimeEstimate}{survfit}(fit, survival = 0.5, ...)

\method{survTimeEstimate}{list}(fit, survival = 0.5, ...)

\method{survTimeEstimate}{coxph}(fit, survival = 0.5, newdata = NULL, ...)
}
\arguments{
\item{fit}{A survival fit object}

\item{survival}{the survival probability for which the timepoint is estimated.}

\item{...}{unused}

\item{newdata}{A dataframe containing predictors for which predictors are desired. See \code{\link[survival]{survfit.coxph}}.}
}
\value{
A named list or matrix with elements surv (estimate), lower and upper (confidence interval). The attribute 'survival' is added to the
result and set to the argument survival probability
}
\description{
Get estimate of timepoints for a given survival probability.
}
\section{Methods (by class)}{
\itemize{
\item \code{survRes}: for survRes objects

\item \code{survfit}: for \code{\link[survival]{survfit}} objects.

\item \code{list}: for \code{\link[rms]{survest.cph}} objects.

\item \code{coxph}: A \code{\link[survival]{coxph}} object.
}}

\examples{
library(rms)
fit = cph(Surv(time=time, event=status) ~ age + sex, data=lung, x=TRUE, y=TRUE, surv=TRUE)
sFit = survfit(fit)
sFit2 = survfit(fit, newdata=lung[1:20,])

survTimeEstimate(sFit) #get median survival for all data points in the dataset.
survTimeEstimate(sFit2) #get median survival for patients 1-20 seperately
}
