% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_functions.R
\name{table_na}
\alias{table_na}
\alias{mean_na}
\alias{sd_na}
\alias{var_na}
\alias{median_na}
\alias{quantile_na}
\alias{range_na}
\alias{sum_na}
\alias{min_na}
\alias{max_na}
\alias{pmin_na}
\alias{pmax_na}
\alias{pmax_na.int}
\alias{pmin_na.int}
\title{Ignore \code{NA} during computations}
\description{
Provides a variety of \code{base} and \code{stats} function, whose default implementation returns \code{NA} if any input value is \code{NA}.
For convenience, these utility functions are provided which ignores \code{NA} and executes the original implementations, see table.
\tabular{ll}{
  \code{table_na}            \tab \code{\link[base]{table}}    \cr
  \code{mean_na}             \tab \code{\link[base]{mean}}     \cr
  \code{min_na}              \tab \code{\link[base]{min}}      \cr
  \code{max_na}              \tab \code{\link[base]{max}}      \cr
  \code{pmin_na}             \tab \code{\link[base]{pmin}}     \cr
  \code{pmax_na}             \tab \code{\link[base]{pmax}}     \cr
  \code{pmin_na.int}         \tab \code{\link[base]{pmin.int}} \cr
  \code{pmax_na.int}         \tab \code{\link[base]{pmax.int}} \cr
  \code{sum_na}              \tab \code{\link[base]{sum}}      \cr
  \code{sd_na}               \tab \code{\link[stats]{sd}}      \cr
  \code{var_na}              \tab \code{\link[stats]{var}}     \cr
  \code{median_na}           \tab \code{\link[stats]{median}}  \cr
  \code{quantile_na}         \tab \code{\link[stats]{quantile}}\cr
  \code{range_na}            \tab \code{\link[base]{range}}    \cr
}
}
\examples{
#create a list with random numbers and 25 NA values
table_na(c(floor(runif(1000,min=0,max=10)), rep(NA,25)))

mean_na(c(1,5,NA,6)) #returns 4
sd_na(c(1,5,NA,6)) #returns 2.645751
var_na(c(1,5,NA,6)) #returns 7
median_na(c(1,5,NA,6)) #returns 5
quantile_na(c(1,5,NA,6))
sum_na(c(1,5,NA,6)) #returns 12
min_na(c(1,5,NA,6)) #returns 1
max_na(c(1,5,NA,6)) #returns 6
pmin_na(1:5, 2) #returns c(1,2,2,2,2)
pmax_na(1:5, 2) #returns c(2,2,3,4,5)
range_na(c(1,5,NA,6)) #returns c(1,6)
}
\seealso{
\code{\link{rm_na}} to remove \code{NA} values from a vector.
}
