% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_animations.R
\name{update.progressbar}
\alias{update.progressbar}
\title{Updates a progress bar}
\usage{
\method{update}{progressbar}(object, progress, ...)
}
\arguments{
\item{object}{A progress bar.}

\item{progress}{Either the proportion progress the bar should be set at, or the current iteration number if \code{n_iterations} is not \code{NULL}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Updates a progress bar
}
