% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{update_additional_args}
\alias{update_additional_args}
\title{Update default parameters with ellipsis}
\usage{
update_additional_args(default, ...)
}
\arguments{
\item{default}{A named list of default values for parameters}

\item{...}{Optional parameters to override the default parameters. A warning is shown if the names do not match.}
}
\value{
The updated list of parameters with possible new values.
}
\description{
Update default parameters with ellipsis
}
\examples{
foo = function(...) {
default = list(a=1)
updated_args = update_additional_args(default, ...)
}
}
