% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-utils.R
\name{update_description}
\alias{update_description}
\title{Create or update a DESCRIPTION field}
\usage{
update_description(fieldname, value, after = NULL)
}
\arguments{
\item{fieldname}{the name of the field}

\item{value}{the new value}

\item{after}{if the field name is new, the name of the field after which the element is placed.}
}
\description{
Create or update a DESCRIPTION field
}
