% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{update_settings}
\alias{update_settings}
\title{Update default function settings}
\usage{
update_settings(default, ...)
}
\arguments{
\item{default}{A named list of default values for settings.}

\item{...}{Optional settings. to override the default settings.}
}
\value{
The updated list of settings. with possible new values.
}
\description{
Uses ellipsis parameter to update a list of default settings.
}
\examples{
foo = function(...) {
  default = list(a=1)
  update_settings(default, ...)
}

foo(a=2, b=3)
}
