% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{use_common_packages}
\alias{use_common_packages}
\alias{list_common_packages}
\title{Load commonly used packages}
\usage{
use_common_packages(...)

list_common_packages()
}
\arguments{
\item{...}{Arguments passed on to \code{use_packages}
\describe{
  \item{load_packages}{Whether to load the selected packages}
  \item{install_packages}{Whether to install the selected packages}
  \item{force_install}{Whether to install packages even if they are installed already}
}}
}
\description{
Utility function to load and install frequently used packages.
}
\details{
\code{load_common_packages} is a wrapper function for \code{\link{use_packages}}. The supplied package names
are available from \code{list_common_packages}.
}
\seealso{
Other developer functions: \code{\link{generic_implementations}},
  \code{\link{set_package_imports}},
  \code{\link{update_settings}},
  \code{\link{use_packages}}, \code{\link{valid_pkgname}}
}
\concept{developer functions}
