% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{valid_pkgname}
\alias{valid_pkgname}
\title{Validate a package name}
\usage{
valid_pkgname(pkg)
}
\arguments{
\item{pkg}{The name of a package. Can be a vector of strings with size of at least 1.
If missing, the function returns the regex to validate the names.}
}
\value{
A boolean indicating whether the package name is valid.
}
\description{
Naming rule obtained from \emph{'Writing R Extensions'} manual.
}
\examples{
valid_pkgname("hgutils") #valid
valid_pkgname("ggplot2") #valid
valid_pkgname("pkg2.-1") #invalid
}
\references{
\href{https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#The-DESCRIPTION-file}{'Writing R Extensions'} manual.
}
\seealso{
Other developer functions: \code{\link{generic_implementations}},
  \code{\link{set_package_imports}},
  \code{\link{update_settings}},
  \code{\link{use_common_packages}},
  \code{\link{use_packages}}
}
\concept{developer functions}
