% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{vec_compareVersion}
\alias{vec_compareVersion}
\title{Compares two vectors of version numbers}
\usage{
vec_compareVersion(a, b)
}
\arguments{
\item{a}{Character strings representing package version numbers.}

\item{b}{Character strings representing package version numbers.}
}
\value{
\code{0} if the numbers are equal, \code{-1} if \code{b} is later
  and \code{1} if \code{a} is later (analogous to the C function
  \code{strcmp}).
}
\description{
Vectorised version of \code{\link[utils]{compareVersion}}.
}
\details{
\R package version numbers are of the form \code{x.y-z} for integers
  \code{x}, \code{y} and \code{z}, with components after \code{x}
  optionally missing (in which case the version number is older than
  those with the components present).
}
\examples{
vec_compareVersion(a=c("1.0","7.2-0"), b=c("1.0-1","7.1-12"))
}
\seealso{
\code{\link[utils]{compareVersion}}
}
