% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-utils.R
\name{write.description}
\alias{write.description}
\title{Writes the DESCRIPTION file}
\usage{
write.description(description)
}
\arguments{
\item{description}{the DESCRIPTION file}
}
\description{
Writes the DESCRIPTION file
}
\details{
The 'Depends', 'Imports' and 'Suggests' fields are sorted before writing the DESCRIPTION file.
}
\examples{
\dontrun{write.description(read.description)}
}
