% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{.get_examples}
\alias{.get_examples}
\title{Displays sample functions for various genres of functions.}
\usage{
.get_examples(genre)
}
\arguments{
\item{genre}{The genre obtained from hgutils::pkg_genre}
}
\description{
Displays sample functions for various genres of functions.
}
