% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{frmt}
\alias{frmt}
\title{Creates a nice string representation of a variable.}
\usage{
frmt(x, show_class = FALSE)
}
\arguments{
\item{x}{The variable for which a string representation is created.}

\item{show_class}{Whether to show the class of 'x'. Defaults to FALSE.}
}
\value{
A character vector with the string representation of 'x'.
}
\description{
Creates a nice string representation of a variable.
}
\examples{
frmt(c(1,2,3))
}
