% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{generic_implementations}
\alias{generic_implementations}
\title{Retrieves generic function implementations}
\usage{
generic_implementations(generic, ...)
}
\arguments{
\item{generic}{The name of the generic function.}

\item{...}{Optional settings}
}
\value{
A vector with class names for which argument '\code{generic}' has an implementation.
}
\description{
Obtains a list of classes for which the supplied generic function has an implementation.
}
\note{
Removes the default generic implementation
}
\examples{
#get a list of classes which have an implementation for graphics::plot
impls = generic_implementations('plot')

}
\seealso{
Other developer functions: \code{\link{set_package_imports}},
  \code{\link{update_settings}},
  \code{\link{use_common_packages}},
  \code{\link{use_packages}}, \code{\link{valid_pkgname}}
}
\concept{developer functions}
