% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{get_square_grid}
\alias{get_square_grid}
\title{Specifies the size of a grid which is as square as possible to fit N objects}
\usage{
get_square_grid(N)
}
\arguments{
\item{N}{Number of objects}
}
\value{
A named list with elements rows and columns specifying the size of the optimal grid.
}
\description{
Specifies the size of a grid which is as square as possible to fit N objects
}
\examples{
get_square_grid(5)
}
