% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{ggplot_breaks}
\alias{ggplot_breaks}
\title{Nice plotting axis breaks}
\usage{
ggplot_breaks(...)
}
\arguments{
\item{...}{Arguments passed on to \code{get_breaks}
\describe{
  \item{limits}{The limits of the axis. May be a vector of 2 elements with lower and upper bounds, or a
single digit (the upperbound, the lowerbound is then assumed to be 0).}
  \item{N}{The unit stepsize; The eventual interval size will be multiples of the divisors of N. Defaults to 10.}
  \item{max_breaks}{Maximum amount of steps, defaults to 10.}
  \item{int_only}{Whether only integer divisors of 'N' may be used for interval sizes, defaults to TRUE.}
  \item{strict}{Whether only multiples of N can be used, defaults to FALSE.}
  \item{include_upper}{Whether the resulting upperbound should go past the upper limit. Defaults to TRUE.}
}}
}
\value{
A \code{\link{get_breaks}} function with filled-in parameters which expects limits.
}
\description{
This makes usage easier in \pkg{ggplot2} as the limits may not be always be known before plotting.
}
\examples{
\dontrun{ggplot() + scale_x_continuous(breaks = ggplot_breaks(N=12, max_breaks=15))}
}
\concept{axis break functions}
