% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Installs and loads specified packages}
\usage{
install_load_packages(..., install_packages = TRUE, load_packages = TRUE,
  force_install = FALSE)
}
\arguments{
\item{...}{A list of package names}

\item{install_packages}{Whether to install the selected packages}

\item{load_packages}{Whether to load the selected packages}

\item{force_install}{Whether to install packages even if they are installed already}
}
\description{
Installs and loads specified packages
}
\examples{
\dontrun{
install_load_packages(c('magrittr','dplyr'))
install_load_packages('magrittr','dplyr',install=FALSE)}
}
\seealso{
Other initialization functions: \code{\link{is_registered}},
  \code{\link{load_common_packages}}, \code{\link{startup}}
}
\concept{initialization functions}
