% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Installs and loads specified packages}
\usage{
install_load_packages(packages, install = TRUE)
}
\arguments{
\item{packages}{A character list of package names}

\item{install}{Whether to install packages when missing, defaults to TRUE}
}
\description{
Installs and loads specified packages
}
\examples{
install_load_packages(c("ggplot2","dplyr"))
}
