% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{load_common_packages}
\alias{load_common_packages}
\title{Loads (and optionally install) commonly used packages}
\usage{
load_common_packages(install = TRUE)
}
\arguments{
\item{install}{Whether to install a package if missing}
}
\description{
Loads (and optionally install) commonly used packages
}
\seealso{
Other initialization functions: \code{\link{install_load_packages}},
  \code{\link{is_registered}}, \code{\link{startup}}
}
\concept{initialization functions}
