% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{load_common_packages}
\alias{load_common_packages}
\title{Loads (and optionally install) commonly used packages}
\usage{
load_common_packages(load_packages = TRUE, install_packages = TRUE,
  force_install = FALSE)
}
\arguments{
\item{load_packages}{Whether to load the selected packages}

\item{install_packages}{Whether to install the selected packages}

\item{force_install}{Whether to install packages even if they are installed already}
}
\description{
Loads (and optionally install) commonly used packages
}
\seealso{
Other initialization functions: \code{\link{install_load_packages}},
  \code{\link{is_registered}}, \code{\link{startup}}
}
\concept{initialization functions}
