% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-utils.R
\name{message_colors}
\alias{message_colors}
\alias{.cinfo}
\alias{.cinfo_bg}
\alias{.cwarn}
\alias{.cwarn_bg}
\alias{.chint}
\alias{.chint_bg}
\alias{.csucc}
\alias{.csucc_bg}
\alias{.cfail}
\alias{.cfail_bg}
\alias{.cnumb}
\alias{.cnumb_bg}
\title{Standardized message colors}
\usage{
.cinfo(...)

.cinfo_bg(...)

.cwarn(...)

.cwarn_bg(...)

.chint(...)

.chint_bg(...)

.csucc(...)

.csucc_bg(...)

.cfail(...)

.cfail_bg(...)

.cnumb(...)

.cnumb_bg(...)
}
\arguments{
\item{...}{the text to be colored.}
}
\value{
A function to color text
}
\description{
Standardized message colors
}
