% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-utils.R
\name{message_colors}
\alias{message_colors}
\alias{.cmess}
\alias{.cwarn}
\alias{.csucc}
\alias{.cfail}
\title{Standardized message colors}
\value{
A function to color text
}
\description{
Standardized message colors
}
\section{Functions}{
\itemize{
\item \code{.cmess}: For warning messages.

\item \code{.cwarn}: For warning messages.

\item \code{.csucc}: For success messages.

\item \code{.cfail}: For fail messages.
}}

