% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_genre.R
\docType{data}
\name{pkg_genre}
\alias{pkg_genre}
\title{List of categorized packages}
\format{A data frame with variables \code{name} and \code{packages}.
The possible genre names are:
\itemize{
\item \code{"data_import"}
\item \code{"image_import"}
\item \code{"ggplot"}
\item \code{"grid"}
\item \code{"survival"}
\item \code{"processing"}
\item \code{"shiny"}
\item \code{"development"}
}}
\usage{
pkg_genre
}
\description{
List of categorized packages
}
\keyword{datasets}
