% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_animations.R
\name{progressbar}
\alias{progressbar}
\alias{update.progressbar}
\alias{render}
\alias{render.progressbar}
\title{Creates an animated progress bar}
\usage{
progressbar(format = "[[|][|/-\\\\][ ]]", width = 25, refresh = 200,
  n_iterations = NULL)

\method{update}{progressbar}(object, progress_ratio = NULL,
  progress_iter = NULL, ...)

render(object, ...)

\method{render}{progressbar}(object, show_percentage = FALSE,
  show_iteration = FALSE, ...)
}
\arguments{
\item{format}{character vector containing the format of the animation. See 'details' for more information.}

\item{width}{progress bar width.}

\item{refresh}{refresh rate in milliseconds of the animation.}

\item{n_iterations}{optional parameter, specifies the number of total iterations. When updating the progress bar it
is then sufficient to specify the current iteration number.}

\item{object}{animated progress bar.}

\item{progress_ratio}{proportion progress the progress bar should be set at.}

\item{progress_iter}{current iteration number if \code{n_iterations} is not \code{NULL}.}

\item{...}{further arguments passed to or from other methods.}

\item{show_percentage}{whether to show percentage progress to the right of the progress bar.}

\item{show_iteration}{whether to show progress as iteration number to the right of the progress bar.}
}
\description{
Creates an animated progress bar
}
\details{
The format of the progress bar is given by a character vector. It consists of 5 parts:
\enumerate{
  \item the left border of the progress bar consisting of 0 or more characters.
  \item a pair of square brackets containing a single character which represents the loaded area.
  \item a pair of square brackets containing 0 or more characters. These are animated on the border between the loaded and unloaded area.
  \item a pair of square brackets containing a single character which represents the unloaded area.
  \item the right border of the progress bar consisting of 0 or more characters.
}
The format follows the following regular expression: \code{^.*?[.?][.*?][.?].*$}
}
\examples{
\dontrun{
# simple progressbar
bar = progressbar(format = "[[|][|/-\\\\][ ]]")
# fancy progressbar using UTF-8 codes
bar2 = progressbar(format="\\u25ba[\\u2589][\\u2580\\u2584][\\u3000]\\u25c4")
n_operations = 1000

for(i in 1:n_operations) {
  bar = update(bar, i/n_operations)
  cat("\\r", render(bar))
  Sys.sleep(0.01)
}}
}
