% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_animations.R
\name{render}
\alias{render}
\title{Render an animated object}
\usage{
render(x, ...)
}
\arguments{
\item{x}{An animated object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector.
}
\description{
Render an animated object
}
