% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_functions.R
\name{rm_na}
\alias{rm_na}
\title{Remove \code{NA}}
\usage{
rm_na(x)
}
\arguments{
\item{x}{A vector which may contain \code{NA} elements.}
}
\value{
A vector without \code{NA} elements.
}
\description{
Removes any \code{NA} element from a vector.
}
\examples{
rm_na(c(1,NA,5,6))
}
