% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{rmNA}
\alias{rmNA}
\title{Removes any NA from a vector}
\usage{
rmNA(vec)
}
\arguments{
\item{vec}{A vector which may contain NA elements}
}
\value{
A vector without NA elements
}
\description{
Removes any NA from a vector
}
\examples{
rmNA(c(1,NA,5,6))
}
