% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{rmNA}
\alias{rmNA}
\title{Removes any NA from a list}
\usage{
rmNA(LIST)
}
\arguments{
\item{LIST}{A list which may contain NA elements}
}
\value{
A list without NA elements
}
\description{
Removes any NA from a list
}
\examples{
rmNA(c(1,NA,5,6))
}
