% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{set_package_imports}
\alias{set_package_imports}
\title{Creates a list of necessary imports for the DESCRIPTION file.}
\usage{
set_package_imports(...)
}
\arguments{
\item{...}{Additional parameters.}
}
\description{
Creates a list of necessary imports for the DESCRIPTION file.
}
\examples{
\dontrun{set_package_imports(skip_prompt=TRUE)}
}
