% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{set_package_imports}
\alias{set_package_imports}
\title{Creates a list of necessary imports for the DESCRIPTION file.}
\usage{
set_package_imports(...)
}
\arguments{
\item{...}{Additional parameters.}
}
\description{
Creates a list of necessary imports for the DESCRIPTION file.
}
\examples{
\dontrun{set_package_imports(skip_prompt=TRUE)}
}
\seealso{
Other developer functions: \code{\link{generic_implementations}},
  \code{\link{update_settings}},
  \code{\link{use_common_packages}},
  \code{\link{use_packages}}, \code{\link{valid_pkgname}}
}
\concept{developer functions}
