% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGUtils.R
\name{startup}
\alias{startup}
\title{Clears the workspace and sets the working directory to specified folder.}
\usage{
startup(folder = NULL)
}
\arguments{
\item{folder}{String to specify folder name. Registered users can indicate subfolders, others needs to specify complete
working directory paths. Use \code{is_registered()} to check whether user registration is enabled}
}
\description{
Clears the workspace and sets the working directory to specified folder.
}
\examples{
startup("source_webinterface/")
}
