% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{startup}
\alias{startup}
\title{Cleans R for use}
\usage{
startup(folder = NULL, rm = TRUE)
}
\arguments{
\item{folder}{String to specify folder name. Registered users can indicate subfolders, others needs to specify complete
working directory paths. Use \code{is_registered()} to check whether user registration is enabled.
If \code{NULL}, the working directory is not changed.}

\item{rm}{Whether to remove objects in the current environment. Defaults to \code{TRUE}.}
}
\description{
Clears workspace, deletes all objects from global environment, clears graphics and (optionally) sets working directory.
}
\examples{
\dontrun{startup('project_folder/')}
}
\concept{initialization functions}
