% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgutils.R
\name{startup}
\alias{startup}
\title{Cleans R for use}
\usage{
startup(clean = TRUE, folder = NULL, verbose = TRUE)
}
\arguments{
\item{clean}{whether to remove objects in the global environment, run garbage collection and to clear graphics. Defaults to \code{TRUE}.}

\item{folder}{folder name to set the current working directory.}

\item{verbose}{whether to print informative messages during cleaning.}
}
\description{
Clears workspace, deletes all objects from global environment, clears graphics and (optionally) sets working directory.
}
\examples{
\dontrun{startup()}
}
\concept{initialization functions}
