% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-estimate.R
\name{time_estimate}
\alias{time_estimate}
\alias{time_estimate.survRes}
\alias{time_estimate.survfit}
\alias{time_estimate.list}
\alias{time_estimate.coxph}
\title{Time point estimation}
\usage{
time_estimate(fit, survival, ...)

\method{time_estimate}{survRes}(fit, survival = 0.5, ...)

\method{time_estimate}{survfit}(fit, survival = 0.5, ...)

\method{time_estimate}{list}(fit, survival = 0.5, ...)

\method{time_estimate}{coxph}(fit, survival = 0.5, newdata = NULL, ...)
}
\arguments{
\item{fit}{\code{\link[=coxph]{survival fit}} object.}

\item{survival}{vector of survival probabilities for which the time points are estimated.}

\item{...}{further arguments passed to or from other methods.}

\item{newdata}{\code{data.frame} containing predictors for which predictors are desired. See \code{\link[survival]{survfit.coxph}}.}
}
\value{
A named list or matrix with elements \code{surv} (estimate), \code{lower} and \code{upper} (confidence interval).
The attribute \code{survival} is added to the result and set to the argument survival probability.
}
\description{
Get estimate of time points for a given survival probability.
}
\section{Methods (by class)}{
\itemize{
\item \code{survRes}: for \code{survRes} objects.

\item \code{survfit}: for \code{\link[survival]{survfit}} objects.

\item \code{list}: for \code{\link[rms]{survest.cph}} objects.

\item \code{coxph}: A \code{\link[survival]{coxph}} object.
}}

\examples{
\dontrun{
fit = rms::cph(Surv(time=time, event=status) ~ age + sex,
  data=lung, x=TRUE, y=TRUE, surv=TRUE)

#get population time estimates
pop = time_estimate(fit)

#get median survival for patients 1-20 seperately.
sub = time_estimate(fit, survival=seq(0.1,0.9,0.1), newdata=lung[1:20,])

fit2 = rms::cph(Surv(time=time, event=status) ~ age + sex,
  data=lung, x=FALSE, y=FALSE, surv=TRUE) #no x nor y
pop2 = time_estimate(fit2, survival=seq(0.1,0.9,0.01)) #get cox.ph estimate
}
}
