% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{update_settings}
\alias{update_settings}
\title{Update default function settings}
\usage{
update_settings(default, ...)
}
\arguments{
\item{default}{A named list of default values for settings.}

\item{...}{Optional settings to override the default settings.}
}
\value{
The updated list of settings with possible new values.
}
\description{
Uses ellipsis parameter to update a list of default settings.
}
\examples{
foo = function(...) {
  default = list(a=1)
  settings = update_settings(default, ...)
}

\dontrun{foo(a=2, b=3)}
}
\seealso{
Other developer functions: \code{\link{generic_implementations}},
  \code{\link{set_package_imports}},
  \code{\link{use_common_packages}},
  \code{\link{use_packages}}, \code{\link{valid_pkgname}}
}
\concept{developer functions}
