% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{use_common_packages}
\alias{use_common_packages}
\title{Load commonly used packages}
\usage{
use_common_packages(...)
}
\arguments{
\item{...}{Arguments passed on to \code{use_packages}
\describe{
  \item{load_packages}{Whether to load the selected packages}
  \item{install_packages}{Whether to install the selected packages}
  \item{force_install}{Whether to install packages even if they are installed already}
}}
}
\description{
\code{load_common_packages} loads (and optionally install) commonly used packages.
A list of used packages is available from \code{common_packages}.
}
\seealso{
Other initialization functions: \code{\link{is_registered}},
  \code{\link{is_valid_pkgname}}, \code{\link{startup}},
  \code{\link{use_packages}}
}
\concept{initialization functions}
