% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{use_packages}
\alias{use_packages}
\title{Installs and loads specified packages}
\usage{
use_packages(..., install_packages = TRUE, load_packages = TRUE,
  force_install = FALSE)
}
\arguments{
\item{...}{List of package names}

\item{install_packages}{Whether to install the selected packages}

\item{load_packages}{Whether to load the selected packages}

\item{force_install}{Whether to install packages even if they are installed already}
}
\description{
Installs and loads specified packages
}
\examples{
\dontrun{
use_packages(c('magrittr','dplyr'))
use_packages('magrittr','dplyr',install_packages=FALSE)}
}
\seealso{
Other initialization functions: \code{\link{is_registered}},
  \code{\link{is_valid_pkgname}}, \code{\link{startup}},
  \code{\link{use_common_packages}}
}
\concept{initialization functions}
