% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{use_packages}
\alias{use_packages}
\title{Prepare packages for usage}
\usage{
use_packages(..., install_packages = TRUE, load_packages = TRUE,
  force_install = FALSE, upgrade = FALSE)
}
\arguments{
\item{...}{List of package names and other options.}

\item{install_packages}{Whether to install the selected packages}

\item{load_packages}{Whether to load the selected packages}

\item{force_install}{Whether to install packages even if they are installed already}

\item{upgrade}{Whether to upgrade outdated packages. Defaults to FALSE.}
}
\description{
Utility function to load and optionally install packages if they are missing. When the function terminates,
packages are installed (if necessary), upgraded to the latest version (if necessary) and loaded.
}
\examples{
\dontrun{
#package names can be given as a vector or one-by-one
use_packages(c('magrittr','dplyr'))
use_packages('magrittr','dplyr',install_packages=FALSE)}
}
\seealso{
Other developer functions: \code{\link{generic_implementations}},
  \code{\link{set_package_imports}},
  \code{\link{update_settings}},
  \code{\link{use_common_packages}},
  \code{\link{valid_pkgname}}
}
\concept{developer functions}
